+++ lib/authority/abilities.rb
@@ -10,11 +10,10 @@ module Authority
   module Abilities
     extend ActiveSupport::Concern
 
-    # Let the Foo model know that its authorizer is called 'FooAuthorizer'
-    # (but let the user change that)
+    # Assume authorizer is `ApplicationAuthorizer` (but let the user change that)
     included do
       class_attribute :authorizer_name
-      self.authorizer_name = "#{name}Authorizer"
+      self.authorizer_name = "ApplicationAuthorizer"
     end
 
     module ClassMethods