+++ lib/generators/authority/install_generator.rb
@@ -8,21 +8,14 @@ module Authority
       desc "Creates an Authority initializer for your application." 
 
       def do_all
+        create_authorizers_directory
+        copy_application_authorizer
         copy_initializer
         copy_forbidden
-        create_authorizers_directory
         message = <<-RUBY
 
         Install complete! See the README on Github for instructions on getting your
         app running with Authority.
-
-        One note: each model needs to know the name of its its authorizer class. 
-        You can specify that in the model like `authorizer_name FooAuthorizer`.
-        If you don't, the `Article` model (for example) will look for `ArticleAuthorizer`.
-
-        To generate one authorizer like that for each of your models, see
-        `rails g authority:authorizers`. If you also want to specify your own
-        parent class for them, use `rails g authority:authorizers MyClass`.
 
         RUBY
         puts message.strip_heredoc
@@ -31,17 +24,21 @@ module Authority
 
       private
 
+      def create_authorizers_directory
+        # creates empty directory if none; doesn't empty the directory
+        empty_directory "app/authorizers" 
+      end
+
+      def copy_application_authorizer
+        template "application_authorizer.rb", "app/authorizers/application_authorizer.rb"
+      end
+
       def copy_initializer
         template "authority_initializer.rb", "config/initializers/authority.rb"
       end
 
       def copy_forbidden
         template "403.html", "public/403.html"
-      end
-
-      def create_authorizers_directory
-        # creates empty directory if none; doesn't empty the directory
-        empty_directory "app/authorizers" 
       end
 
     end