+++ spec/support/no_authorizer_model.rb
@@ -1,5 +1,0 @@
-# No corresponding Authorizer is defined for this model
-
-class NoAuthorizerModel
-  include Authority::Abilities
-end